UPSTREAM: mwifiex: cancel cmd timer and free curr_cmd in shutdown process
commit 084c7189acb3f969c855536166042e27f5dd703f upstream.

curr_cmd points to the command that is in processing or waiting
for its command response from firmware. If the function shutdown
happens to occur at this time we should cancel the cmd timer and
put the command back to free queue.

Tested-by: Marco Cesarano <marco@marvell.com>
Signed-off-by: Bing Zhao <bzhao@marvell.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

BUG=chrome-os-partner:40159
TEST=We have ported some critical fixes from wireless-3.8 to
wireless-3.4. With this patch series applied, ran autotests and
no regression found.

Change-Id: I7f43c5c50387b0729e15ccbda9ae00c9531d8dfa
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/277838
Reviewed-by: Rebecca Silberstein <silberst@chromium.org>