s5p-mfc: Fix for mfc-sysmmu clock gate issue
Issue is seen on Ares where MFC clock has to be turned ON
when accessing MFC-SYSMMU-L and R registers. This patch
removes the hack where MFC clock was turned ON always. Now
the clock is turned ON and OFF whenever the sysmmu access
is taking place.

BUG=chrome-os-partner:19203
TEST=Tested with MFC decoder & encoder test app [1]
and vdau using secomx, on peach-pit

[1]http://git.infradead.org/users/kmpark/public-apps

Change-Id: Ia6c44ac403278898a8e1d086136d570bc49958e1
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Signed-off-by: kiran AVND <avnd.kiran@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/56212
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>