BACKPORT: drm/i915: enable VT switchless resume v3
With the other bits in place, we can do this safely.

v2: disable backlight on suspend to prevent premature enablement on resume
v3: disable CRTCs on suspend to allow RTD3 (Kristen)

This regresses resume time on 3.8 systems, but this is because it causes
the panel timings to not be ignored.

BUG=chromium:577441
TEST=suspend/resume on peppy/link

Signed-off-by: Jesse Barnes <jbarnes@virtuousgeek.org>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@gmail.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 24576d23976746cb52e7700c4cadbf4bc1bc3472)
Conflicts:
	drivers/gpu/drm/i915/i915_drv.c
	drivers/gpu/drm/i915/intel_fb.c
	Minor conflicts in i915_drv.c. intel_fb.c was renamed to intel_fbdev.c
Signed-off-by: Derek Basehore <dbasehore@chromium.org>

Change-Id: I33b7a70c773a062fda6978197bf3ba9103b77876
Reviewed-on: https://chromium-review.googlesource.com/322641
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>