UPSTREAM: ARM: tegra: keep power on to SD slot on Dalmore
Set "regulator-always-on" for the SD slot on Dalmore, so that SD cards
work. This used to work, since this regulator is on by default, but was
broken by commit "ARM: tegra: dalmore: add TPS65090 node", since that
didn't specify always-on for this regulator.

In the long run, the regulators should all be hooked up to the SDHCI
device nodes. However, we haven't done that for any of the Tegra boards
yet, so to be consistent, this patch simply forces the regulator on,
rather than hooking it up and making it work differently to other boards.

Reported-by: Rhyland Klein <rklein@nvidia.com>
Acked-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Stephen Warren <swarren@nvidia.com>
(cherry picked from commit 15d5ef4d9a1f6fd7bcf706e41f13d5029a91760f)
(cherry-picked from:
	git://git.kernel.org/pub/scm/linux/kernel/git/swarren/linux-tegra.git)
Signed-off-by: Olof Johansson <olofj@chromium.org>