UPSTREAM: Don't disable hpet emulation on suspend
(This syncs up our early commit with the upstream version of the same).

There's a bug where rtc alarms are ignored after the rtc cmos suspends but
before the system finishes suspend. Since hpet emulation is disabled and it
still handles the interrupts, a wake event is never registered which is done
from the rtc layer. This reverts an earlier commit which disables hpet
emulation. To fix the problem mentioned in that commit, the hpet_rtc_timer_init
function is called directly on resume.

BUG=chromium:221857
TEST=suspend_stress_test and inserting delays in suspend path with an rtc alarm

Change-Id: I7969efd864b2d79802072e8167f764e680e3d2d3
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48836