UPSTREAM: clk: Fixup locking issues for clk_set_parent
Updating the clock tree topology must be protected with the spinlock
when doing clk_set_parent, otherwise we can not handle the migration
of the enable_count in a safe manner.

While issuing the .set_parent callback to make the clk-hw perform the
switch to the new parent, we can not hold the spinlock since it is must
be allowed to be slow path. This complicates error handling, but is still
possible to achieve.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Cc: Rajagopal Venkat <rajagopal.venkat@linaro.org>
Signed-off-by: Mike Turquette <mturquette@linaro.org>
(cherry picked from commit a68de8e4ab2756d8e125cf327f044c895b6a6b3d)

BUG=chrome-os-partner:23164
TEST=No regressions noticed on Pit or Snow

Change-Id: Id087a92d7803553ad4f593c5d6fa41c8e08e8eeb
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/172187
Reviewed-by: Doug Anderson <dianders@chromium.org>