UPSTREAM: ARM: tegra: add power gpio keys to DT
This adds the power gpio key to DT and enable the wakeup of the gpio key
for the device. The Seaboard and paz00 already had the power gpio key
binding and the power key of Whistler was on KBC. So these boards' device
tree didn't include in this patch.

Signed-off-by: Joseph Lo <josephl@nvidia.com>
Signed-off-by: Stephen Warren <swarren@nvidia.com>
(cherry picked from commit 5741a2560afa66dce9e53752888241310d8e7c93)
Signed-off-by: Olof Johansson <olofj@chromium.org>