CHROMIUM: clocksource: tegra: enable arch_timer
Architected timer is the local timer for Cortex-A15. Adding the support
for Tegra.

Note:
This patch was origianlly upstreamed for Kernel 3.8 but after talking
with "Rob Herring" who is moving the arch local timer to use CLKSRC_OF
init. That means we don't need this patch after Rob's patch is merged (It
was merged in Kernel 3.10). The link of the talking was below:
http://www.mail-archive.com/linux-tegra@vger.kernel.org/msg10034.html

BUG=none
TEST=checking the Architected local timer is running and system switching
     to timer-based delay loop

Change-Id: I94dbf9e62394f5fdce40cd91371515fd6111ec59
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/64161
Commit-Queue: Olof Johansson <olofj@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>