HACK: ARM: dts: Turn on all 32K clocks on cros5250 boards
The 32K clocks coming out of the Maxim PMIC should be always on for
cros5250 boards.  They go to the AP, the TPM, and the WiFi module and
those won't work so well without these clocks.

For some reason we had these clocks specified as "regulator-boot-on"
in 3.4.  That doesn't seem right and also doesn't seem to work in 3.8.
Switch them to "regulator-always-on" since that seems more correct and
actually works.

I've listed this as a "HACK" since we'll want to move away from
specifing these as regulators once we can switch to the common clock
framework for exynos and can use the upstream 77686 clock driver.

BUG=chrome-os-partner:17354
TEST=Boot kernel-next on snow and show that WiFi actually scans now
(it still won't associate, but that's believed to be a different
problem).
TEST=Boot and look for 32k clocks in /sys/class/regulators.  Ensure
all are enabled.
TEST='i2cget -f -y 0 9 0x7f b' now gives back 0xf.

Change-Id: Ica9b4ed409f537f742d1fbf43e4d72851c29e7d8
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47702