temp: mmc: dw_mmc: exynos: Turn SDIO interrupts on
The mmc part in exynos supports SDIO interrupts and they work fine, so
turn the capability on.  With this I see download speeds increase
about 10x.  This is a port of a change present in the Chrome OS tree
that can be found at <https://gerrit.chromium.org/gerrit/#/c/26729/>.

(This patch is a copy of one I posted up to LKML at
https://patchwork.kernel.org/patch/2429661/).

BUG=chrome-os-partner:17354
TEST=Download a 500MB file and see average download speeds go from
~600k/s to ~6000k/s as reported by curl.  Download still works OK.

Change-Id: I4bfd59ae9a35bed0f9f9f579de6a174eece1538f
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
[sonnyrao: 3.8: re-landing after pulling in upstream changes]
Reviewed-on: https://chromium-review.googlesource.com/173523
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>