temp: mmc: dw_mmc: exynos: Turn SDIO interrupts on
The mmc part in exynos supports SDIO interrupts and they work fine, so
turn the capability on.  With this I see download speeds increase
about 10x.  This is a port of a change present in the Chrome OS tree
that can be found at <https://gerrit.chromium.org/gerrit/#/c/26729/>.

(This patch is a copy of one I posted up to LKML at
https://patchwork.kernel.org/patch/2429661/).

BUG=chrome-os-partner:17354
TEST=Download a 500MB file and see average download speeds go from
~600k/s to ~6000k/s as reported by curl.  Download still works OK.

Signed-off-by: Doug Anderson <dianders@chromium.org>
Change-Id: I481812cbe31a8dbf9e0ad7f8f6867a2ec91d1262
Reviewed-on: https://gerrit.chromium.org/gerrit/47865
Reviewed-by: Paul Stewart <pstew@chromium.org>