HACK: mwifiex: disable uap0 and p2p0
There is a known issue in firmware where it may get stuck if the
Network Manager tries to scan on p2p0 (or uap0) interface in some
scenarios.

Until that's solved, hack it out.

This patch was provided by Bing Zhao from Marvell and attached to the
referenced bug.

BUG=chrome-os-partner:17354
TEST=Boot on Snow and see that WiFi no longer gives a timeout when
failing to associate.  Repeat 3 times.
TEST=Associate to WiFi (GoogleGuest) and download 500MB file.  Repeat
3 times.

Change-Id: I008e91f7f4395556677dc5358b9cb9eeb26f2bdf
Signed-off-by: Bing Zhao <bzhao@marvell.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47863
Reviewed-by: Paul Stewart <pstew@chromium.org>