BACKPORT: drm/i915: Scale ring, rather than ia, frequency on Haswell
Haswell introduces a separate frequency domain for the ring (uncore). So
where we used to increase the CPU (IA) clock with GPU busyness, we now
need to scale the ring frequency directly instead. As the ring limits
our memory bandwidth, it is vital for performance that when the GPU is
busy, we increase the frequency of the ring to increase the available
memory bandwidth.

v2: Fix the algorithm to actually use the scaled gpu frequency for the ring.
v3: s/max_ring_freq/min_ring_freq/ as that is what it is

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Jesse Barnes <jbarnes@virtuousgeek.org>
Reviewed-by: Jesse Barnes <jbarnes@virtuousgeek.org>
[danvet: Add space checkpatch complained about.]
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 3ebecd07d382c022e93c560c56114eec1d6d2cdd)

BUG=None
TEST=This patch will boost graphical performance in Haswell, with the
     situation of memory intense high GPU usage and less of CPU.
     Because its hard to run heavy tests suite like phoronix
     in chrome-OS, the performance increase is validated by considering
     FPS as the test parameter on webgl experiment satisfying the
     above usage.

Change-Id: Ib8293aa92c9bed82246ecb2573b80dd279723933
Reviewed-on: https://gerrit.chromium.org/gerrit/63072
Tested-by: Josh Triplett <josh@joshtriplett.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Josh Triplett <josh@joshtriplett.org>