UPSTREAM: wireless: regulatory: fix channel disabling race condition
Upstream commit 990de49f74e772b6db5208457b7aa712a5f4db86.

When a full scan 2.4 and 5 GHz scan is scheduled, but then the 2.4 GHz
part of the scan disables a 5.2 GHz channel due to, e.g. receiving
country or frequency information, that 5.2 GHz channel might already
be in the list of channels to scan next. Then, when the driver checks
if it should do a passive scan, that will return false and attempt an
active scan. This is not only wrong but can also lead to the iwlwifi
device firmware crashing since it checks regulatory as well.

Fix this by not setting the channel flags to just disabled but rather
OR'ing in the disabled flag. That way, even if the race happens, the
channel will be scanned passively which is still (mostly) correct.

BUG=chrome-os-partner:20691
TEST=tested that connectivity works over ath9k and iwl7000

Change-Id: If739fd0947b2c1e7a6dffd5af5987b0921dee94b
Cc: stable@vger.kernel.org
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/59780
Reviewed-by: Paul Stewart <pstew@chromium.org>